+++ docs/resources/releases.md
@@ -64,12 +64,12 @@ The details for configuration are available in the [YugabyteDB Anywhere Upgrade 
 
 - `chart_path` (String) File path where the release helm chart is stored.
 - `file_path` (String) File path where the release binary is stored.
-- `gcs` (Block List, Max: 1) Location of release binary in GCS. (see [below for nested schema](#nestedblock--gcs))
+- `gcs` (Block List, Max: 1, Deprecated) Location of release binary in GCS.Deprecated since YugabyteDB Anywhere 2.20. Please use 'http' instead. (see [below for nested schema](#nestedblock--gcs))
 - `http` (Block List, Max: 1) Location of release binary in HTTP. (see [below for nested schema](#nestedblock--http))
 - `image_tag` (String) Docker Image Tag for the release.
 - `notes` (List of String) Release Notes.
 - `packages` (Block List) Package path and architecture. (see [below for nested schema](#nestedblock--packages))
-- `s3` (Block List, Max: 1) Location of release binary in S3. (see [below for nested schema](#nestedblock--s3))
+- `s3` (Block List, Max: 1, Deprecated) Location of release binary in S3. Deprecated since YugabyteDB Anywhere 2.20. Please use 'http' instead. (see [below for nested schema](#nestedblock--s3))
 - `state` (String) State of Release.
 - `timeouts` (Block, Optional) (see [below for nested schema](#nestedblock--timeouts))